Fix unlink/xattr deadlock
Gunnar Beutner nicely summerized the root cause of this issue
in a comment for Github issue #266.

The problem here is that prune_icache() tries to evict/delete
both the xattr directory inode as well as at least one xattr
inode contained in that directory. Here's what happens:

1. File is created.
2. xattr is created for that file (behind the scenes a xattr
   directory and a file in that xattr directory are created)
3. File is deleted.
4. Both the xattr directory inode and at least one xattr
   inode from that directory are evicted by prune_icache();
   prune_icache() acquires a lock on both inodes before it
   calls ->evict() on the inodes

When the xattr directory inode is evicted zfs_zinactive attempts
to delete the xattr files contained in that directory. While
enumerating these files zfs_zget() is called to obtain a reference
to the xattr file znode - which tries to lock the xattr inode.
However that very same xattr inode was already locked by
prune_icache() further up the call stack, thus leading to a
deadlock.

This can be reliably reproduced like this:

---

This patch proposes to fix the issue by deferring the call to
zfs_rmnode() until after the inodes are evicted.  This is
accomplished by dispatching the zfs_rmnode() call to the existing
zfs_iput_taskq.  There is no requirement that it must complete
before zfs_zinactive() returns.  The inode is already unlinked
from the namespace an inactive so this should be safe.

Issue #266